split out dialogs into separate files, fix icon alignment
