+++ libs/ui-toolkit/src/components/icon/icon.tsx
@@ -14,6 +14,7 @@ export const Icon = ({ size = 16, name, className }: IconProps) => {
   const effectiveClassName = classNames(
     'inline-block',
     'fill-current',
+    'align-text-bottom',
     {
       'w-20': size === 20,
       'h-20': size === 20,
